Hoist XRouteCalculation to its own file
Step 2 of the Big Break Up, simply because it was next in the original
file, after RouteCalculation.
