+++ PyRoute/Calculation/TradeCalculation.py
@@ -64,231 +64,6 @@ class OwnedWorldCalculation(RouteCalculation):
         return weight
 
 
-class XRouteCalculation(RouteCalculation):
-    distance_weight = [0, 95, 90, 85, 80, 75, 70]
-    capSec_weight = [0, 95, 90, 85, 75, 75, 70]
-    inSec_weight = [0, 140, 110, 85, 70, 95, 140]
-    impt_weight = [0, 90, 80, 70, 70, 110, 140]
-
-    def __init__(self, galaxy):
-        super(XRouteCalculation, self).__init__(galaxy)
-        self.route_reuse = 5
-
-    def base_range_routes(self, star, neighbor):
-        pass
-
-    def base_route_filter(self, star, neighbor):
-        if not AllyGen.are_allies(star.alg_code, neighbor.alg_code):
-            return True
-        if not AllyGen.imperial_align(star.alg_code):
-            return True
-        if star.zone in ['R', 'F'] or neighbor.zone in ['R', 'F']:
-            return True
-
-        return False
-
-    def generate_routes(self):
-        self.distance_weight = self.capSec_weight
-        self.generate_base_routes()
-        self.capital = [star for star in self.galaxy.ranges if \
-                        AllyGen.imperial_align(star.alg_code) and star.tradeCode.other_capital]
-        self.secCapitals = [star for star in self.galaxy.ranges if \
-                            AllyGen.imperial_align(star.alg_code) and star.tradeCode.sector_capital]
-        self.subCapitals = [star for star in self.galaxy.ranges if \
-                            AllyGen.imperial_align(star.alg_code) and star.tradeCode.subsector_captial]
-
-    def routes_pass_1(self):
-        # Pass 1: Get routes at J6  Capital and sector capitals 
-        # This should be 1 element long
-        self.logger.info(self.capital)
-        if len(self.capital) == 0:
-            return
-        for star in self.secCapitals:
-            self.get_route_between(self.capital[0], star, self.calc_trade(25), Star.heuristicDistance)
-
-        for star in self.secCapitals:
-            localCapital = {'coreward': None, 'spinward': None, 'trailing': None, 'rimward': None,
-                            'corespin': None, 'coretrail': None, 'rimspin': None, 'rimtrail': None}
-
-            if star.sector.coreward:
-                localCapital['coreward'] = self.find_sector_capital(star.sector.coreward)
-                if localCapital['coreward'] and localCapital['coreward'].sector.spinward:
-                    localCapital['corespin'] = self.find_sector_capital(localCapital['coreward'].sector.spinward)
-                if localCapital['coreward'] and localCapital['coreward'].sector.trailing:
-                    localCapital['coretrail'] = self.find_sector_capital(localCapital['coreward'].sector.trailing)
-            if star.sector.rimward:
-                localCapital['rimward'] = self.find_sector_capital(star.sector.rimward)
-                if localCapital['rimward'] and localCapital['rimward'].sector.spinward:
-                    localCapital['rimspin'] = self.find_sector_capital(localCapital['rimward'].sector.spinward)
-                if localCapital['rimward'] and localCapital['rimward'].sector.trailing:
-                    localCapital['rimtrail'] = self.find_sector_capital(localCapital['rimward'].sector.trailing)
-            if star.sector.spinward:
-                localCapital['spinward'] = self.find_sector_capital(star.sector.spinward)
-                if localCapital['spinward'] and localCapital['spinward'].sector.coreward and not localCapital[
-                    'corespin']:
-                    localCapital['corespin'] = self.find_sector_capital(localCapital['spinward'].sector.coreward)
-                if localCapital['spinward'] and localCapital['spinward'].sector.rimward and not localCapital['rimspin']:
-                    localCapital['rimwspin'] = self.find_sector_capital(localCapital['spinward'].sector.rimward)
-            if star.sector.trailing:
-                localCapital['trailing'] = self.find_sector_capital(star.sector.trailing)
-                if localCapital['trailing'] and localCapital['trailing'].sector.coreward and not localCapital[
-                    'coretrail']:
-                    localCapital['coretrail'] = self.find_sector_capital(localCapital['trailing'].sector.coreward)
-                if localCapital['trailing'] and localCapital['trailing'].sector.rimward and not localCapital[
-                    'rimtrail']:
-                    localCapital['rimtrail'] = self.find_sector_capital(localCapital['trailing'].sector.rimward)
-
-            for neighbor in localCapital.values():
-                if neighbor and not self.galaxy.ranges.has_edge(star, neighbor):
-                    self.get_route_between(star, neighbor, self.calc_trade(25), Star.heuristicDistance)
-
-    def routes_pass_2(self):
-        # Step 2a - re-weight the routes to be more weighted to J4 than J6
-        self.reweight_routes(self.inSec_weight)
-
-        secCapitals = self.secCapitals + self.capital
-        for sector in self.galaxy.sectors.values():
-
-            secCap = [star for star in secCapitals if star.sector == sector]
-            self.logger.info(secCap)
-            subCap = [star for star in self.subCapitals if star.sector == sector]
-
-            if len(secCap) == 0:
-                if len(subCap) == 0:
-                    continue
-                else:
-                    self.logger.info("{} has subsector capitals but no sector capital".format(sector.name))
-                    for star in subCap:
-                        capital = self.find_nearest_capital(star, secCapitals)
-                        self.get_route_between(capital[0], star, self.calc_trade(23), Star.heuristicDistance)
-            else:
-                for star in subCap:
-                    if self.galaxy.ranges.has_edge(secCap[0], star):
-                        continue
-                    self.get_route_between(secCap[0], star, self.calc_trade(23), Star.heuristicDistance)
-
-        for star in self.subCapitals:
-            routes = [neighbor for neighbor in self.subCapitals if \
-                      neighbor != star and neighbor.hex_distance(star) <= 40]
-            for neighbor in routes:
-                self.get_route_between(star, neighbor, self.calc_trade(23), Star.heuristicDistance)
-
-    def routes_pass_3(self):
-        self.reweight_routes(self.impt_weight)
-        important = [star for star in self.galaxy.ranges if \
-                     AllyGen.imperial_align(star.alg_code) and star.tradeCount == 0
-                     and (star.importance >= 4 or 'D' in star.baseCode or 'W' in star.baseCode)]
-
-        jumpStations = [star for star in self.galaxy.ranges if star.tradeCount > 0]
-
-        self.logger.info('Important worlds: {}, jump stations: {}'.format(len(important), len(jumpStations)))
-
-        important2 = []
-        for star in important:
-            if star in jumpStations:
-                continue
-            for neighbor in self.galaxy.stars.neighbors_iter(star):
-                if star.hex_distance(neighbor) > 4:
-                    continue
-                if neighbor in jumpStations:
-                    self.get_route_between(star, neighbor, self.calc_trade(21), None)
-                    jumpStations.append(star)
-            if star.tradeCount == 0:
-                important2.append(star)
-
-        important3 = []
-        for star in important2:
-            for neighbor in self.galaxy.stars.neighbors_iter(star):
-                if neighbor in jumpStations:
-                    self.get_route_between(star, neighbor, self.calc_trade(21), None)
-                    jumpStations.append(star)
-            if star.tradeCount == 0:
-                important3.append(star)
-                self.logger.info("No route for important world: {}".format(star))
-
-        capitalList = self.capital + self.secCapitals + self.subCapitals
-        for star in important3:
-            capital = self.find_nearest_capital(star, capitalList)
-            self.get_route_between(capital[0], star, self.calc_trade(21), None)
-
-    def calculate_routes(self):
-        self.logger.info('XRoute pass 1')
-        self.routes_pass_1()
-
-        self.logger.info('XRoute pass 2')
-        self.routes_pass_2()
-
-        self.logger.info('XRoute pass 3')
-        self.routes_pass_3()
-
-    def reweight_routes(self, weightList):
-        self.distance_weight = weightList
-        for (star, neighbor, data) in self.galaxy.stars.edges(data=True):
-            data['weight'] = self.route_weight(star, neighbor)
-            for _ in range(1, min(data['count'], 5)):
-                data['weight'] -= data['weight'] // self.route_reuse
-
-    def find_nearest_capital(self, world, capitals):
-        dist = (None, 9999)
-        for capital in capitals:
-            newDist = capital.hex_distance(world)
-            if newDist < dist[1]:
-                dist = (capital, newDist)
-        return dist
-
-    def find_sector_capital(self, sector):
-        for world in self.secCapitals:
-            if world.sector == sector:
-                return world
-        return None
-
-    def get_route_between(self, star, target, trade, heuristic):
-        try:
-            route = nx.astar_path(self.galaxy.stars, star, target, heuristic)
-        except nx.NetworkXNoPath:
-            return
-        self.galaxy.ranges.add_edge(star, target, distance=star.hex_distance(target))
-
-        distance = 0
-        start = route[0]
-        start.tradeCount += 1
-        for end in route[1:]:
-            distance += start.hex_distance(end)
-            end.tradeCount += 1
-            data = self.galaxy.stars[start][end]
-            data['trade'] = max(trade, data[start][end]['trade'])
-            data['count'] += 1
-            data['weight'] -= (data['weight'] - data['distance']) // self.route_reuse
-            start = end
-
-        self.galaxy.ranges[route[0]][route[-1]]['actual distance'] = distance
-        self.galaxy.ranges[route[0]][route[-1]]['jumps'] = len(route) - 1
-
-    def route_weight(self, star, target):
-        dist = star.hex_distance(target)
-        weight = self.distance_weight[dist]
-        if star.port in 'CDEX':
-            weight += 25
-        if star.port in 'DEX':
-            weight += 25
-        if star.zone in 'RF' or target.zone in 'RF':
-            weight += 50
-        if star.popCode == 0 or target.popCode == 0:
-            weight += 25
-        weight -= 3 * (star.importance + target.importance)
-        weight -= 6 if 'S' in star.baseCode or 'S' in target.baseCode else 0
-        weight -= 6 if 'W' in star.baseCode or 'W' in target.baseCode else 0
-        weight -= 3 if 'N' in star.baseCode or 'N' in target.baseCode else 0
-        weight -= 3 if 'D' in star.baseCode or 'D' in target.baseCode else 0
-        weight -= 6 if star.tradeCode.subsecor_capital or target.tradeCode.subsector_capital else 0
-        weight -= 6 if star.tradeCode.other_capital or target.tradeCode.other_capital else 0
-        weight -= 6 if star.tradeCode.sector_capital or target.tradeCode.sector_capital else 0
-        assert 0 < weight, "Weight of edge between " + str(star) + " and " + str(
-            target) + " must be positive"
-
-        return weight
-
-
 class TradeCalculation(RouteCalculation):
     """
     Perform the trade calculations by generating the routes
